requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pycparser==2.22
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1

django-allauth==0.61.1
dj-rest-auth==6.0.0